Rearrange state to use globalstate
